Trial for improvement on new logger : move class Color and struct LevelInfo to a new source file
