+++ src/libs/antares/logs/CMakeLists.txt
@@ -1,6 +1,7 @@
 set(PROJ logs)
 set(HEADERS
         include/antares/${PROJ}/logs.h
+        include/antares/${PROJ}/logger-utils.h
         include/antares/${PROJ}/hostinfo.h
         include/antares/${PROJ}/hostname.hxx
 )
